requirements.txt
Django==1.6
Markdown==2.4
South==0.8.1
argparse==1.2.1
dj-database-url==0.2.1
gunicorn==0.17.4
psycopg2==2.5
wsgiref==0.1.2
ez-setup==0.9
smartypants==1.8.3
titlecase==0.5.1
typogrify==2.0.3
django-recaptcha-works==0.3.4
